Move i03 dcm to different folder
