+++ src/dodal/beamlines/i18.py
@@ -12,7 +12,7 @@ from dodal.common.visit import (
     LocalDirectoryServiceClient,
     StaticVisitPathProvider,
 )
-from dodal.devices.dcm import DCM
+from dodal.devices.i03.dcm import DCM
 from dodal.devices.i18.diode import Diode
 from dodal.devices.i18.KBMirror import KBMirror
 from dodal.devices.i18.table import Table
