+++ src/dodal/plans/verify_undulator_gap.py
@@ -2,7 +2,7 @@ from typing import Protocol, runtime_checkable
 
 from bluesky import plan_stubs as bps
 
-from dodal.devices.dcm import DCM
+from dodal.devices.i03.dcm import DCM
 from dodal.devices.undulator import Undulator
