+++ tests/plans/conftest.py
@@ -7,7 +7,7 @@ from ophyd_async.core import PathProvider, StandardDetector, init_devices
 from ophyd_async.sim import PatternGenerator, SimBlobDetector, SimMotor
 from tests.constants import UNDULATOR_ID_GAP_LOOKUP_TABLE_PATH
 
-from dodal.devices.dcm import DCM
+from dodal.devices.i03.dcm import DCM
 from dodal.devices.undulator import Undulator
